+++ src/interrogate/coverage.py
@@ -26,13 +26,11 @@ class BaseInterrogateResult:
         classes, methods, and functions).
     :attr int covered: number of objects covered by docstrings.
     :attr int missing: number of objects not covered by docstrings.
-    :attr int skipped: number of modules skipped.
     """
 
     total = attr.ib(init=False, default=0)
     covered = attr.ib(init=False, default=0)
     missing = attr.ib(init=False, default=0)
-    skipped = attr.ib(init=False, default=0)
 
     @property
     def perc_covered(self):
@@ -60,21 +58,20 @@ class InterrogateFileResult(BaseInterrogateResult):
 
     filename = attr.ib(default=None)
     ignore_module = attr.ib(default=False)
-    visitor = attr.ib(repr=False, default=None)
+    nodes = attr.ib(repr=False, default=None)
 
     def combine(self):
         """Tally results from each AST node visited."""
-        for node in self.visitor.graph.nodes:
+        # for node in self.visitor.graph.nodes:
+        for node in self.nodes:
             if node.node_type == "Module":
                 if self.ignore_module:
-                    self.skipped += 1
                     continue
             self.total += 1
             if node.covered:
                 self.covered += 1
 
         self.missing = self.total - self.covered
-        self.skipped = self.visitor.skipped
 
 
 @attr.s
@@ -96,7 +93,6 @@ class InterrogateResults(BaseInterrogateResult):
             self.covered += result.covered
             self.missing += result.missing
             self.total += result.total
-            self.skipped += result.skipped
 
 
 class InterrogateCoverage:
@@ -168,6 +164,13 @@ class InterrogateCoverage:
         self.common_base = utils.get_common_base(filenames)
         return filenames
 
+    def _filter_nodes(self, nodes):
+        """Remove empty modules when ignoring modules."""
+        is_empty = 1 == len(nodes)
+        if is_empty and self.config.ignore_module:
+            return []
+        return nodes
+
     def _get_file_coverage(self, filename):
         """Get coverage results for a particular file."""
         with open(filename, "r", encoding="utf-8") as f:
@@ -177,10 +180,14 @@ class InterrogateCoverage:
         visitor = visit.CoverageVisitor(filename=filename, config=self.config)
         visitor.visit(parsed_tree)
 
+        filtered_nodes = self._filter_nodes(visitor.graph.nodes)
+        if len(filtered_nodes) == 0:
+            return
+
         results = InterrogateFileResult(
             filename=filename,
             ignore_module=self.config.ignore_module,
-            visitor=visitor,
+            nodes=filtered_nodes,
         )
         results.combine()
         return results
@@ -190,7 +197,9 @@ class InterrogateCoverage:
         results = InterrogateResults()
         file_results = []
         for f in filenames:
-            file_results.append(self._get_file_coverage(f))
+            result = self._get_file_coverage(f)
+            if result:
+                file_results.append(result)
         results.file_results = file_results
 
         results.combine()
@@ -207,14 +216,17 @@ class InterrogateCoverage:
 
     def _get_detailed_row(self, node, filename):
         """Generate a row of data for the detailed view."""
-        padding = "  " * node.level
-
         filename = filename[len(self.common_base) + 1 :]
-        name = "{} (module)".format(filename)
-        if node.node_type != "Module":
+
+        if node.node_type == "Module":
+            if self.config.ignore_module:
+                return [filename, ""]
+            name = "{} (module)".format(filename)
+        else:
             name = node.path.split(":")[-1]
             name = "{} (L{})".format(name, node.lineno)
 
+        padding = "  " * node.level
         name = "{}{}".format(padding, name)
         status = "MISSED" if not node.covered else "COVERED"
         return [name, status]
@@ -239,7 +251,7 @@ class InterrogateCoverage:
         verbose_tbl.append(header)
         verbose_tbl.append(utils.TABLE_SEPARATOR)
         for file_result in combined_results.file_results:
-            nodes = file_result.visitor.graph.nodes
+            nodes = file_result.nodes
             nodes = sorted(nodes, key=_sort_nodes)
             for n in nodes:
                 verbose_tbl.append(