+++ src/interrogate/visit.py
@@ -43,7 +43,6 @@ class CoverageVisitor(ast.NodeVisitor):
         self.stack = []
         self.graph = nx.Graph()
         self.config = config
-        self.skipped = 0
 
     @staticmethod
     def _has_doc(node):
@@ -161,7 +160,6 @@ class CoverageVisitor(ast.NodeVisitor):
         :param ast.ClassDef node: a class AST node.
         """
         if self._is_class_ignored(node):
-            self.skipped += 1
             return
         self._visit_helper(node)
 
@@ -171,7 +169,6 @@ class CoverageVisitor(ast.NodeVisitor):
         :param ast.FunctionDef node: a function/method AST node.
         """
         if self._is_func_ignored(node):
-            self.skipped += 1
             return
         self._visit_helper(node)
 
@@ -181,6 +178,5 @@ class CoverageVisitor(ast.NodeVisitor):
         :param ast.AsyncFunctionDef node: a async function/method AST node.
         """
         if self._is_func_ignored(node):
-            self.skipped += 1
             return
         self._visit_helper(node)